Files Uploaded for Breast Cancer Survival
The files contains my work towards being a part of python data analytics world. I have tried to study the survival of breast cancer patients with respect to the age at which the disease was detected, No of Aux nodes available and its impact on survival, When  they were operated and if the operation was successful or not.